Remove ez_setup
ez_setup used only for installing of the setuptools, and it is available on
every up to date linux distribution. It is bad thing to force the major part of
users to install the latest setuptools version required by ez_setup, just to
provide "a little bit easier" install for the minor part.